+++ exercisesClass01/src/main/java/cl/sebastian_alejandro/talentofuturoexercises/class01/blocktwo/Block2Exercise01.java
@@ -1,0 +1,24 @@
+/*
+ * This file is part of TalentoFuturoExercises.
+ *
+ * Copyright 2024 Sebestar
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * See the LICENSE file for more information.
+ */
+
+package cl.sebastian_alejandro.talentofuturoexercises.class01.blocktwo;
+
+public class Block2Exercise01 {
+
+  // [EN] Create a function for each of the exercises in block 1
+  // [ES] Cree una función para cada uno de los ejercicios del bloque 1
+
+  public static void main(String[] args) {
+
+    // SKIPPED EXERCISE / EJERCICIO SALTADO
+    // [EN] Block 1 algorithms will be transferred to grouped algorithm repositories
+    // [ES] Los algoritmos del bloque 1 serán traspasados a repositorios de algoritmos agrupados
+    System.out.println("Skipped Exercise 01 in Block 2 -> Good bye!");
+  }
+}